feat(api): created all necessary POST routes for Bibliomania, separated files and routes to import to server.js
